requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clip==1.0
contourpy==1.3.2
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
decorator==5.2.1
decord==0.6.0
diffusers==0.37.0.dev0
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.132.0
ffmpy==1.0.0
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gradio==6.6.0
gradio_client==2.1.0
groovy==0.1.2
grpcio==1.78.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lightning-utilities==0.15.3
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.1
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
open_clip_torch==3.2.0
opencv-python==4.13.0.92
orjson==3.11.7
packaging==25.0
pandas==2.3.3
photomaker==1.0.1
pillow==11.3.0
pip==26.0.1
proglog==0.1.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytorch-lightning==2.6.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
timm==1.0.25
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchmetrics==1.8.2
torchvision==0.25.0
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
wheel==0.46.3
xformers==0.0.35
yarl==1.22.0
zipp==3.23.0
